requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
diffusers @ git+https://github.com/huggingface/diffusers.git@500b9cf184d9a6aa950ccbb9a4d967b95877196d
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.0.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nvitop==1.5.3
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentry-sdk==2.41.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spark-parser==1.9.0
stack-data==0.6.3
sympy==1.14.0
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uncompyle6==3.9.3
urllib3==2.5.0
wandb==0.22.2
wcwidth==0.2.14
xdis==6.1.8
zipp==3.23.0
